+++ src/commands/ReviseCommand.ts
@@ -1,10 +1,5 @@
-import chalk from "chalk"
 import { Command, Option } from "clipanion"
-import ora from "ora"
-import { input } from "@inquirer/prompts"
 import GenerativeCommand from "../lib/GenerativeCommand"
-import { config } from "../config"
-import { reviseCommandStream } from "../services/reviseCommand"
 
 export default class ReviseCommand extends GenerativeCommand {
   static usage = Command.Usage({
@@ -18,24 +13,10 @@ export default class ReviseCommand extends GenerativeCommand {
 
   static paths = [["revise"], ["r"]]
 
-  command = Option.String({ required: true })
+  command = Option.Rest({ required: 1 })
 
   async execute() {
     this.assertGeminiKey()
-
-    const revisePrompt = await input({
-      message: "Enter your revision",
-    })
-    const spinner = ora().start()
-
-    const cmd = this.command
-    const ci = config.customInstructions
-    const query = "[null]"
-    const cmdStream = await reviseCommandStream(query, cmd, revisePrompt, ci)
-    spinner.stop()
-
-    const revisedCmd = await this.writeStream(cmdStream, chunk => chalk.cyan(chunk))
-
-    this.respond(query, revisedCmd)
+    await this.revise(this.command.join(" "))
   }
 }