+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Added Gemini finish reason handling
+
 ## [0.3.0] - 2024-04-16
 
 ### Added