+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Added Gemini finish reason handling
+- Implemented exit on cancel
 
 ## [0.3.0] - 2024-04-16